+++ reqwest-middleware/src/middleware.rs
@@ -1,6 +1,5 @@
+use http::Extensions;
 use reqwest::{Client, Request, Response};
-use std::sync::Arc;
-use task_local_extensions::Extensions;
 
 use crate::error::{Error, Result};
 
@@ -12,7 +11,7 @@ use crate::error::{Error, Result};
 /// ```
 /// use reqwest::{Client, Request, Response};
 /// use reqwest_middleware::{ClientBuilder, Middleware, Next, Result};
-/// use task_local_extensions::Extensions;
+/// use http::Extensions;
 ///
 /// struct TransparentMiddleware;
 ///
@@ -74,7 +73,7 @@ where
 #[derive(Clone)]
 pub struct Next<'a> {
     client: &'a Client,
-    middlewares: &'a [Arc<dyn Middleware>],
+    middlewares: &'a [Box<dyn Middleware>],
 }
 
 #[cfg(not(target_arch = "wasm32"))]
@@ -83,7 +82,7 @@ pub type BoxFuture<'a, T> = std::pin::Pin<Box<dyn std::future::Future<Output = T
 pub type BoxFuture<'a, T> = std::pin::Pin<Box<dyn std::future::Future<Output = T> + 'a>>;
 
 impl<'a> Next<'a> {
-    pub(crate) fn new(client: &'a Client, middlewares: &'a [Arc<dyn Middleware>]) -> Self {
+    pub(crate) fn new(client: &'a Client, middlewares: &'a [Box<dyn Middleware>]) -> Self {
         Next {
             client,
             middlewares,
@@ -97,7 +96,7 @@ impl<'a> Next<'a> {
     ) -> BoxFuture<'a, Result<Response>> {
         if let Some((current, rest)) = self.middlewares.split_first() {
             self.middlewares = rest;
-            Box::pin(current.handle(req, extensions, self))
+            current.handle(req, extensions, self)
         } else {
             Box::pin(async move { self.client.execute(req).await.map_err(Error::from) })
         }