+++ reqwest-tracing/src/reqwest_otel_span_builder.rs
@@ -4,7 +4,7 @@ use matchit::Router;
 use reqwest::header::{HeaderMap, HeaderValue};
 use reqwest::{Request, Response, StatusCode as RequestStatusCode, Url};
 use reqwest_middleware::{Error, Result};
-use task_local_extensions::Extensions;
+use http::Extensions;
 use tracing::{warn, Span};
 
 use crate::reqwest_otel_span;